+++ appmap/test/test_recording.py
@@ -3,6 +3,7 @@ import json
 import os
 import sys
 
+from appmap._implementation.recording import Recorder, wrap_exec_module
 from .appmap_test_base import AppMapTestBase
 
 
@@ -35,3 +36,26 @@ class TestRecording(AppMapTestBase):
         generated_appmap = self.normalize_appmap(appmap.generation.dump(r))
 
         assert generated_appmap == expected_appmap
+
+    def test_exec_module_protection(self, monkeypatch):
+        """
+        Test that recording.wrap_exec_module properly protects against
+        rewrapping a wrapped exec_module function.  Repeatedly wrap
+        the function, up to the recursion limit, then call the wrapped
+        function.  If wrapping protection is working properly, there
+        won't be a problem.  If wrapping protection is broken, this
+        test will fail with a RecursionError.
+        """
+
+        def exec_module():
+            pass
+
+        def do_import(*args, **kwargs):  # pylint: disable=unused-argument
+            pass
+        monkeypatch.setattr(Recorder, 'do_import', do_import)
+        f = exec_module
+        for _ in range(sys.getrecursionlimit()):
+            f = wrap_exec_module(f)
+
+        f()
+        assert True